move packets around
