+++ src/main/java/at/hannibal2/skyhanni/data/hypixel/chat/event/GuildChatEvent.kt
@@ -1,7 +1,5 @@
 package at.hannibal2.skyhanni.data.hypixel.chat.event
 
-import at.hannibal2.skyhanni.data.hypixel.chat.AbstractChatEvent
-
 class GuildChatEvent(
     author: String,
     message: String,
