+++ src/main/java/at/hannibal2/skyhanni/data/hypixel/chat/event/NpcChatEvent.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.data.hypixel.chat
+package at.hannibal2.skyhanni.data.hypixel.chat.event
 
 class NpcChatEvent(
     author: String,
